Refactor: Moved XStreamXmlPlannerBenchmarkFactoryTest to NQueens*, because of IntelliJ tests failing (wrong working dir)
