--- optaplanner-benchmark/src/test/resources/org/optaplanner/benchmark/impl/testdataInputSolutionFile.xml
@@ -1,8 +1,0 @@
-<TestdataSolution id="1">
-  <entityList id="2">
-    <TestdataEntity id="3"/>
-  </entityList>
-  <valueList id="4">
-    <TestdataValue id="5"/>
-  </valueList>
-</TestdataSolution>
