--- optaplanner-benchmark/src/test/resources/org/optaplanner/benchmark/impl/testdataPlannerBenchmarkConfigXStream.xml
@@ -1,20 +1,0 @@
-<plannerBenchmark xStreamId="1">
-  <solverBenchmark xStreamId="2">
-    <name>Default Solver Benchmark</name>
-    <solver xStreamId="3">
-      <solutionClass>org.optaplanner.core.impl.testdata.domain.TestdataSolution</solutionClass>
-      <entityClass>org.optaplanner.core.impl.testdata.domain.TestdataEntity</entityClass>
-      <scoreDirectorFactory xStreamId="4">
-        <scoreDefinitionType>SIMPLE</scoreDefinitionType>
-        <scoreDrl>org/optaplanner/core/api/solver/testdataScoreRules.drl</scoreDrl>
-      </scoreDirectorFactory>
-      <termination xStreamId="5">
-        <secondsSpentLimit>1</secondsSpentLimit>
-      </termination>
-    </solver>
-    <problemBenchmarks xStreamId="6">
-      <xStreamAnnotatedClass>org.optaplanner.core.impl.testdata.domain.TestdataSolution</xStreamAnnotatedClass>
-      <inputSolutionFile>src/test/resources/org/optaplanner/benchmark/impl/testdataInputSolutionFile.xml</inputSolutionFile>
-    </problemBenchmarks>
-  </solverBenchmark>
-</plannerBenchmark>
