--- optaplanner-benchmark/src/test/resources/org/optaplanner/benchmark/impl/testdataPlannerBenchmarkConfigXStreamInherited.xml
@@ -1,23 +1,0 @@
-<plannerBenchmark xStreamId="1">
-  <inheritedSolverBenchmark xStreamId="2">
-    <subSingleCount>2</subSingleCount>
-  </inheritedSolverBenchmark>
-  <solverBenchmark xStreamId="3">
-    <name>Default Solver Benchmark</name>
-    <solver xStreamId="4">
-      <solutionClass>org.optaplanner.core.impl.testdata.domain.TestdataSolution</solutionClass>
-      <entityClass>org.optaplanner.core.impl.testdata.domain.TestdataEntity</entityClass>
-      <scoreDirectorFactory xStreamId="5">
-        <scoreDefinitionType>SIMPLE</scoreDefinitionType>
-        <scoreDrl>org/optaplanner/core/api/solver/testdataScoreRules.drl</scoreDrl>
-      </scoreDirectorFactory>
-      <termination xStreamId="6">
-        <secondsSpentLimit>1</secondsSpentLimit>
-      </termination>
-    </solver>
-    <problemBenchmarks xStreamId="7">
-      <xStreamAnnotatedClass>org.optaplanner.core.impl.testdata.domain.TestdataSolution</xStreamAnnotatedClass>
-      <inputSolutionFile>src/test/resources/org/optaplanner/benchmark/impl/testdataInputSolutionFile.xml</inputSolutionFile>
-    </problemBenchmarks>
-  </solverBenchmark>
-</plannerBenchmark>
