+++ optaplanner-examples/src/test/resources/org/optaplanner/examples/nqueens/persistence/nqueensSimpleBenchmarkConfig.xml
@@ -1,0 +1,25 @@
+<plannerBenchmark xStreamId="1">
+  <benchmarkDirectory>local/data/nqueens</benchmarkDirectory>
+  <solverBenchmark xStreamId="2">
+    <name>Entity Tabu Search</name>
+    <solver xStreamId="3">
+      <constructionHeuristic xStreamId="4">
+        <constructionHeuristicType>FIRST_FIT_DECREASING</constructionHeuristicType>
+      </constructionHeuristic>
+      <localSearch xStreamId="5">
+        <changeMoveSelector xStreamId="6">
+          <selectionOrder>ORIGINAL</selectionOrder>
+        </changeMoveSelector>
+        <acceptor xStreamId="7">
+          <entityTabuSize>5</entityTabuSize>
+        </acceptor>
+        <forager xStreamId="8"/>
+      </localSearch>
+    </solver>
+    <problemBenchmarks xStreamId="9">
+      <xStreamAnnotatedClass>org.optaplanner.examples.nqueens.domain.NQueens</xStreamAnnotatedClass>
+      <inputSolutionFile>data/nqueens/unsolved/32queens.xml</inputSolutionFile>
+    </problemBenchmarks>
+    <subSingleCount>2</subSingleCount>
+  </solverBenchmark>
+</plannerBenchmark>
